refactor: mdx-to-markdown directory structure
